Name of commit, then should be a detailed description
An F-test is any statistical test in which the test statistic has an F-distribution under the null hypothesis. It is most often used when comparing statistical models that have been fitted to a data set, in order to identify the model that best fits the population
from which the data were sampled. Exact F-tests mainly arise when the models have been
fitted to the data using least squares. The name was coined by George W. Snedecor, in
honour of Ronald Fisher. Fisher initially developed the statistic as the variance ratio in
the 1920s.